requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# DFL Framework
# p2pfl>=0.1.0  # Uncomment when using P2PFL

# Utilities
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Code Quality
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0